requirements.txt
Django==2.1.5
django-bootstrap-form==3.4
djangorestframework==3.9.1
html5lib==1.0.1
Pillow==5.4.1
PyPDF2==1.26.0
pytz==2018.9
reportlab==3.5.13
six==1.12.0
webencodings==0.5.1
xhtml2pdf==0.2.3
